requirements.txt
pillow
numpy
matplotlib
lxml_html_clean
scipy
scikit-learn
pandas
tensorflow<=2.17.0
#tensorflow-macos>=2.12.1; sys_platform == 'darwin'
tensorflow-metal>=0.8; sys_platform == 'darwin'
tensorflow-model-optimization
tf_keras
opencv-python
notebook
jupyter